removed CustomerFactory from example1
