+++ crabzilla-kotlin-example1/crabzilla-kotlin-example1-service/src/main/java/io/github/crabzilla/example1/Example1Launcher.kt
@@ -5,7 +5,7 @@ import com.google.inject.Inject
 import io.github.crabzilla.example1.customer.*
 import io.github.crabzilla.vertx.entity.EntityCommandExecution
 import io.github.crabzilla.vertx.entity.EntityCommandHandlerVerticle
-import io.github.crabzilla.vertx.entity.EntityCommandHttpRpcVerticle
+import io.github.crabzilla.vertx.entity.EntityCommandRestVerticle
 import io.github.crabzilla.vertx.helpers.ConfigHelper.cfgOptions
 import io.github.crabzilla.vertx.helpers.StringHelper
 import io.github.crabzilla.vertx.projection.EventsProjectionVerticle
@@ -27,7 +27,7 @@ class Example1Launcher {
   internal lateinit var projectionVerticle: EventsProjectionVerticle<CustomerSummaryDao>
 
   @Inject
-  internal lateinit var httpRpcVerticle: EntityCommandHttpRpcVerticle<Customer>
+  internal lateinit var restVerticle: EntityCommandRestVerticle<Customer>
 
   @Inject
   internal lateinit var cmdVerticle: EntityCommandHandlerVerticle<Customer>
@@ -67,11 +67,11 @@ class Example1Launcher {
 
         injector.injectMembers(launcher)
 
-        vertx.deployVerticle(launcher.projectionVerticle) { event -> log.debug("Deployed ? {}", event.succeeded()) }
+        val inDeploymentOrder = listOf(launcher.projectionVerticle, launcher.cmdVerticle, launcher.restVerticle)
 
-        vertx.deployVerticle(launcher.cmdVerticle) { event -> log.debug("Deployed ? {}", event.succeeded()) }
-
-        vertx.deployVerticle(launcher.httpRpcVerticle) { event -> log.debug("Deployed ? {}", event.succeeded()) }
+        for (v in inDeploymentOrder) {
+          vertx.deployVerticle(v) { event -> log.debug("Deployed ? {}", event.succeeded()) }
+        }
 
         // a test
          launcher.justForTest(vertx);
